+++ bephim/src/main/java/org/example/bephim/repository/MovieRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Long> {
+public interface MovieRepository extends JpaRepository<Movie, Integer> {
     @Query("select m from Movie m where m.title like %:title%")
      List<Movie> findByTitle(@Param("title") String title);
 }
